refactor: move useCardSide to custom hooks #37
